+++ batch/src/main/java/kr/co/techcourse/prolog/batch/job/eve/MemberBatchJob.java
@@ -1,7 +1,7 @@
-package kr.co.techcourse.prolog.batch.job;
+package kr.co.techcourse.prolog.batch.job.eve;
 
-import kr.co.techcourse.prolog.batch.domain.Member;
-import kr.co.techcourse.prolog.batch.domain.MemberRepository;
+import kr.co.techcourse.prolog.batch.job.eve.domain.Member;
+import kr.co.techcourse.prolog.batch.job.eve.domain.MemberRepository;
 import org.springframework.batch.core.Job;
 import org.springframework.batch.core.Step;
 import org.springframework.batch.core.configuration.annotation.JobBuilderFactory;
