+++ batch/src/main/java/kr/co/techcourse/prolog/batch/job/eve/StepDataShareBean.java
@@ -1,4 +1,4 @@
-package kr.co.techcourse.prolog.batch.job;
+package kr.co.techcourse.prolog.batch.job.eve;
 
 import java.util.HashMap;
 import java.util.Map;
