+++ batch/src/main/java/kr/co/techcourse/prolog/batch/job/eve/domain/Member.java
@@ -1,4 +1,4 @@
-package kr.co.techcourse.prolog.batch.domain;
+package kr.co.techcourse.prolog.batch.job.eve.domain;
 
 import javax.persistence.Entity;
 import javax.persistence.GeneratedValue;
